refactory layout
