+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Routes } from 'react-router';
 import { Home } from './components/Home';
 import { News } from './components/News';
-import { LayoutBackOffice } from './components/layout/LayoutBackOffice';
+import { LayoutBackOffice } from './layout/LayoutBackOffice';
 import './App.css';
 import Login from './components/Login/Login';
