+++ src/constants/CardLayoutInfo.js
